% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfgbuilder-package.R
\docType{package}
\name{sfgbuilder-package}
\alias{sfgbuilder}
\alias{sfgbuilder-package}
\title{sfgbuilder: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Michael Sumner \email{mdsumner@gmail.com}

}
\keyword{internal}
